functions/file-processor/requirements.txt
google-cloud-storage==2.10.0
functions-framework==3.4.0 
